refactored main.js page to work with the also refactored TablatureGenerator. Removed files that are no longer in use.
