+++ Main.js
@@ -1,39 +1,52 @@
-const NewTablatureGenerator =  require('./NewTablatureGenerator.js')
+const TablatureGenerator =  require('./TablatureGenerator.js')
 const Instrument = require('./Instrument.js')
+let allEighths = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16]
+let allQuarters = [1, 3, 5, 7, 9, 11, 13, 15]
+let guitar = new Instrument(
+  "guitar",
+  ["e", "B", "G", "D", "A", "E"],
+  [
+    {
+      strings: ["E", "A", "D"],
+      rhythm: allQuarters,
+      restChance: .50
+    },
+    {
+      strings: ["G", "B", "e"],
+      rhythm: allEighths,
+      restChance: .50
+    }
+  ]
+)
 
-let guitar = new Instrument("guitar", [3, 3])
+let ukelele = new Instrument(
+  "ukelele",
+  ["A", "E", "C", "g"],
+  [
+    {
+      strings: ["g", "C"],
+      rhythm: allQuarters,
+      restChance: .50
+    },
+    {
+      strings: ["E", "A"],
+      rhythm: allEighths,
+      restChance: .50
+    }
+  ]
+)
 
-let ukelele = new Instrument("ukelele", [1, 1, 2])
+let logTabs = (instrument, measureLength) => {
+  if (instrument === "guitar"){
+    instrument = guitar
+  } else if (instrument === "ukelele"){
+    instrument = ukelele
+  }
 
-let arrOfEmptyStrings
-let lowJazzChordEmptyStringConfig = [1, 5]
-let highJazzChordEmptyStringConfig = [4, 6]
-let jazzChordRandomConfig = Math.random()
-if (jazzChordRandomConfig >= 0.5) {
-  arrOfEmptyStrings = lowJazzChordEmptyStringConfig
-} else {
-  arrOfEmptyStrings = highJazzChordEmptyStringConfig
-}
-
-let jazzGuitar = new Instrument("jazz guitar", [3, 1], arrOfEmptyStrings)
-
-let logJazzGuitarTabs = (measureLength) => {
-  let generator = new TablatureGenerator(jazzGuitar)
-  console.log(generator.generate(measureLength))
-}
-
-let logGuitarTabs = (measureLength) => {
-  let generator = new TablatureGenerator(guitar)
-  console.log(generator.generate(measureLength))
-}
-
-let logUkeleleTabs = (measureLength) => {
-  let generator = new TablatureGenerator(ukelele)
-  console.log(generator.generate(measureLength))
+  let tabs = new TablatureGenerator(instrument)
+  console.log(tabs.generate(measureLength))
 }
 
 module.exports = {
-  logGuitarTabs,
-  logUkeleleTabs,
-  logJazzGuitarTabs
+  logTabs
 }
